refactor with long id

+++ app/it/sf/manager/SessionManager.scala
@@ -17,7 +17,7 @@ object SessionManager extends ApplicationLoggerImpl with UserService with PageVi
 
   //TODO: revisit it
   def logWebPage(webPage: String, request: Request[play.api.mvc.AnyContent]) {
-    val idUser: Option[Int] = SessionManager.getUserSession(request.session) match {
+    val idUser: Option[Long] = SessionManager.getUserSession(request.session) match {
       case None => Some(Defines.DEFAULT_USER_ID)
       case o: Option[Object] =>
         o.getOrElse(Defines.DEFAULT_USER_ID) match {
