+++ app/it/sf/models/Configuration.scala
@@ -7,11 +7,11 @@ import it.sf.models.{InterestTable, UserTable}
 import org.joda.time.DateTime
 import com.github.tototoshi.slick.H2JodaSupport._
 
-case class Configuration(id: Option[Int] = None, key: String, value: String, creationTime: DateTime)
+case class Configuration(id: Option[Long] = None, key: String, value: String, creationTime: DateTime)
 
 class ConfigurationTable(tag: Tag) extends Table[Configuration](tag, "CONFIGURATION") {
 
-  def id = column[Int]("ID", O.PrimaryKey, O.AutoInc)
+  def id = column[Long]("ID", O.PrimaryKey, O.AutoInc)
   def key = column[String]("ID_USER", O.NotNull)
   def value = column[String]("ID_INTEREST", O.NotNull)
   def creationTime = column[DateTime]("CREATION_TIME", O.NotNull)
