+++ app/it/sf/models/OAuth2Info.scala
@@ -8,10 +8,10 @@ import it.sf.service.UserService
 /**
  * Created by diego on 12/05/14.
  */
-case class OAuth2Info(userId: Int, accessToken: String, tokenType: Option[String], expiresIn: Option[Int], refreshToken: Option[String])
+case class OAuth2Info(userId: Long, accessToken: String, tokenType: Option[String], expiresIn: Option[Int], refreshToken: Option[String])
 
 class OAuth2InfoTable(tag: Tag) extends Table[OAuth2Info](tag, "USER_OAUTH2INFO") with UserService {
-  def userId = column[Int]("USER_ID", O.PrimaryKey)
+  def userId = column[Long]("USER_ID", O.PrimaryKey)
 
   def accessToken = column[String]("ACCESS_TOKEN", O.NotNull)
   def tokenType = column[String]("TOKEN_TYPE", O.Nullable)
